+++ src/main/java/zenpad/branch/Branch.java
@@ -1,12 +1,12 @@
 package zenpad.branch;
 
-import zenpad.fileopener.FileOpenerPanel;
+import zenpad.explorer.ExplorerPanel;
 
 /**
  * Branch class represents a branch in the file tree structure.
  * It contains the parent name, display names, and file names associated with the branch.
  * 
- * @see FileOpenerPanel
+ * @see ExplorerPanel
  * @param parentName The name of the parent branch.
  * @param displayNames An array of display names for the files in the branch.
  * @param fileNames An array of file names for the files in the branch.
