+++ src/main/java/zenpad/core/AppFrame.java
@@ -8,7 +8,7 @@ import javax.swing.JFrame;
 import javax.swing.JSplitPane;
 import javax.swing.JTabbedPane;
 
-import zenpad.fileopener.FileOpenerPanel;
+import zenpad.explorer.ExplorerPanel;
 import zenpad.misc.factory.ComponentFactory;
 import zenpad.misc.manager.NoteSyncManager;
 import zenpad.note.NotePanel;
@@ -22,7 +22,7 @@ import zenpad.toolbar.Toolbar;
  * The frame is created with a title and has a default close operation to exit the application.
  * 
  * @see TabManager
- * @see FileOpenerPanel
+ * @see ExplorerPanel
  * @param title The title of the application window.
  */
 
@@ -33,8 +33,8 @@ public class AppFrame extends JFrame {
     // Instance of TabManager object for external tab management logic
     private TabManager tabManager;
 
-    // Instance of FileOpenerPanel object for sidebar management
-    private FileOpenerPanel fileOpenerPanel;
+    // Instance of ExplorerPanel object for sidebar management
+    private ExplorerPanel explorerPanel;
 
     // Instance of CodeRunner object for code execution
     private CodeRunner codeRunner;
@@ -72,7 +72,7 @@ public class AppFrame extends JFrame {
 
         tabManager = new TabManager(tabbedPane, this::updateNotePanelVis);
         codeRunner = new CodeRunner();
-        fileOpenerPanel = new FileOpenerPanel(tabManager, notePanel);
+        explorerPanel = new ExplorerPanel(tabManager, notePanel);
 
         addTabSwitchListener();
     }
@@ -106,7 +106,7 @@ public class AppFrame extends JFrame {
     }
 
     private void setupOuterSplit() {
-        outerSplitPane = LayoutManager.createOuterSplitPane(fileOpenerPanel.getPanel(), innerSplitPane);
+        outerSplitPane = LayoutManager.createOuterSplitPane(explorerPanel.getPanel(), innerSplitPane);
     }
 
     public void updateNotePanelVis() {
